requirements.txt
asgiref==3.5.2
Django==3.2
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
Pillow==9.1.1
PyJWT==1.7.1
pytz==2022.1
sqlparse==0.4.2
